# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: NoorAl Quran Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /bookmarks:
    get:
      operationId: getBookmarks
      tags: [bookmarks]
      summary: List all bookmarks for the authenticated user
      responses:
        "200":
          description: List of bookmarks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Bookmark"
    post:
      operationId: createBookmark
      tags: [bookmarks]
      summary: Bookmark an ayah
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookmarkInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bookmark"
  /bookmarks/{id}:
    delete:
      operationId: deleteBookmark
      tags: [bookmarks]
      summary: Remove a bookmark
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Notes ──────────────────────────────────────────────────────────────────
  /notes:
    get:
      operationId: getNotes
      tags: [notes]
      summary: List all notes for the authenticated user
      parameters:
        - name: surahId
          in: query
          schema:
            type: integer
          description: Filter by surah
      responses:
        "200":
          description: List of notes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Note"
    post:
      operationId: createNote
      tags: [notes]
      summary: Create a note on an ayah
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoteInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
  /notes/{id}:
    patch:
      operationId: updateNote
      tags: [notes]
      summary: Update a note
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoteUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
    delete:
      operationId: deleteNote
      tags: [notes]
      summary: Delete a note
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Hard Ayahs ─────────────────────────────────────────────────────────────
...
  /moods:
    get:
      operationId: getMoods
      tags: [moods]
      summary: Get mood tags for the authenticated user
      parameters:
        - name: surahId
          in: query
          schema:
            type: integer
        - name: ayahNumber
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of mood tags
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AyahMood"
    post:
      operationId: saveMood
      tags: [moods]
      summary: Save or update a mood tag for an ayah
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AyahMoodInput"
      responses:
        "201":
          description: Saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AyahMood"
  /moods/{id}:
    delete:
      operationId: deleteMood
      tags: [moods]
      summary: Remove a mood tag
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Khatm ─────────────────────────────────────────────────────────────────
...
  /pinned-verses:
    get:
      operationId: getPinnedVerses
      tags: [pinned-verses]
      summary: List pinned verses for the authenticated user
      responses:
        "200":
          description: List of pinned verses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PinnedVerse"
    post:
      operationId: pinVerse
      tags: [pinned-verses]
      summary: Pin a verse (max 10)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PinnedVerseInput"
      responses:
        "201":
          description: Pinned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PinnedVerse"
  /pinned-verses/{id}:
    delete:
      operationId: unpinVerse
      tags: [pinned-verses]
      summary: Unpin a verse
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Unpinned

  # ── Collections ────────────────────────────────────────────────────────────
...
  /collections/{id}/verses:
    post:
      operationId: addVerseToCollection
      tags: [collections]
      summary: Add a verse to a collection
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CollectionVerseInput"
      responses:
        "201":
          description: Added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollectionVerse"
        "409":
          description: Already in collection
  /collections/{id}/verses/{verseKey}:
    delete:
      operationId: removeVerseFromCollection
      tags: [collections]
      summary: Remove a verse from a collection
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: verseKey
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Removed

  # ── Donations ──────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Bookmark:
      type: object
      required: [id, userId, surahId, ayahNumber, surahNameEn, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: string
        surahId:
          type: integer
        ayahNumber:
          type: integer
        surahNameEn:
          type: string
        surahNameAr:
          type: string
        ayahText:
          type: string
        createdAt:
          type: string
    BookmarkInput:
      type: object
      required: [surahId, ayahNumber, surahNameEn]
      properties:
        surahId:
          type: integer
        ayahNumber:
          type: integer
        surahNameEn:
          type: string
        surahNameAr:
          type: string
        ayahText:
          type: string

    # ── Notes ─────────────────────────────────────────────────────────────────
    Note:
      type: object
      required: [id, userId, surahId, ayahNumber, content, createdAt, updatedAt]
      properties:
        id:
          type: integer
        userId:
          type: string
        surahId:
          type: integer
        ayahNumber:
          type: integer
        surahNameEn:
          type: string
        content:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
    NoteInput:
      type: object
      required: [surahId, ayahNumber, content]
      properties:
        surahId:
          type: integer
        ayahNumber:
          type: integer
        surahNameEn:
          type: string
        content:
          type: string
    NoteUpdate:
      type: object
      properties:
        content:
          type: string

    # ── Hard Ayahs ────────────────────────────────────────────────────────────
...
    AyahMood:
      type: object
      required: [id, userId, surahId, ayahNumber, mood, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: string
        surahId:
          type: integer
        ayahNumber:
          type: integer
        mood:
          type: string
          enum: [peace, reflection, gratitude, awe, hope]
        createdAt:
          type: string
    AyahMoodInput:
      type: object
      required: [surahId, ayahNumber, mood]
      properties:
        surahId:
          type: integer
        ayahNumber:
          type: integer
        mood:
          type: string
          enum: [peace, reflection, gratitude, awe, hope]

    # ── Khatm ─────────────────────────────────────────────────────────────────
...
    PinnedVerse:
      type: object
      required: [id, userId, surahId, ayahNumber, surahNameEn, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: string
        surahId:
          type: integer
        ayahNumber:
          type: integer
        surahNameEn:
          type: string
        surahNameAr:
          type: ["string", "null"]
        ayahText:
          type: ["string", "null"]
        translation:
          type: ["string", "null"]
        createdAt:
          type: string
    PinnedVerseInput:
      type: object
      required: [surahId, ayahNumber, surahNameEn]
      properties:
        surahId:
          type: integer
        ayahNumber:
          type: integer
        surahNameEn:
          type: string
        surahNameAr:
          type: string
        ayahText:
          type: string
        translation:
          type: string

    # ── Collections ───────────────────────────────────────────────────────────
...
    CollectionVerse:
      type: object
      required: [id, collectionId, userId, verseKey, surahId, ayahNumber, surahNameEn, createdAt]
      properties:
        id:
          type: integer
        collectionId:
          type: integer
        userId:
          type: string
        verseKey:
          type: string
        surahId:
          type: integer
        ayahNumber:
          type: integer
        surahNameEn:
          type: string
        ayahText:
          type: ["string", "null"]
        translation:
          type: ["string", "null"]
        note:
          type: ["string", "null"]
        createdAt:
          type: string
    CollectionVerseInput:
      type: object
      required: [verseKey, surahId, ayahNumber, surahNameEn]
      properties:
        verseKey:
          type: string
        surahId:
          type: integer
        ayahNumber:
          type: integer
        surahNameEn:
          type: string
        ayahText:
          type: string
        translation:
          type: string
        note:
          type: string
...